+++ app/Http/Controllers/Settings/SubscriptionsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Settings;
 use App\Helpers\DateHelper;
 use Illuminate\Http\Request;
 use App\Helpers\InstanceHelper;
+use App\Exceptions\StripeException;
 use App\Http\Controllers\Controller;
 
 class SubscriptionsController extends Controller
@@ -27,10 +28,15 @@ class SubscriptionsController extends Controller
         }
 
         $planId = auth()->user()->account->getSubscribedPlanId();
+        try {
+            $nextBillingDate = auth()->user()->account->getNextBillingDate();
+        } catch (StripeException $e) {
+            $nextBillingDate = trans('app.unknown');
+        }
 
         return view('settings.subscriptions.account', [
             'planInformation' => InstanceHelper::getPlanInformationFromConfig($planId),
-            'nextBillingDate' => auth()->user()->account->getNextBillingDate(),
+            'nextBillingDate' => $nextBillingDate,
         ]);
     }
 
@@ -118,7 +124,7 @@ class SubscriptionsController extends Controller
             return redirect()->route('settings.index');
         }
 
-        auth()->user()->account->subscription(auth()->user()->account->getSubscribedPlanName())->cancelNow();
+        auth()->user()->account->subscriptionCancel();
 
         return redirect()->route('settings.subscriptions.downgrade.success');
     }
@@ -134,40 +140,16 @@ class SubscriptionsController extends Controller
             return redirect()->route('settings.index');
         }
 
-        $stripeToken = $request->input('stripeToken');
-
-        $plan = InstanceHelper::getPlanInformationFromConfig($request->input('plan'));
-        $errorMessage = '';
-
         try {
-            auth()->user()->account->newSubscription($plan['name'], $plan['id'])
-                        ->create($stripeToken, [
-                            'email' => auth()->user()->email,
-                        ]);
-
-            return redirect()->route('settings.subscriptions.upgrade.success');
-        } catch (\Stripe\Error\Card $e) {
-            // Since it's a decline, \Stripe\Error\Card will be caught
-            $body = $e->getJsonBody();
-            $err = $body['error'];
-            $errorMessage = trans('settings.stripe_error_card', ['message' => $err['message']]);
-        } catch (\Stripe\Error\RateLimit $e) {
-            // Too many requests made to the API too quickly
-            $errorMessage = trans('settings.stripe_error_rate_limit');
-        } catch (\Stripe\Error\Authentication $e) {
-            // Authentication with Stripe's API failed
-            // (maybe you changed API keys recently)
-            $errorMessage = trans('settings.stripe_error_authentication');
-        } catch (\Stripe\Error\ApiConnection $e) {
-            // Network communication with Stripe failed
-            $errorMessage = trans('settings.stripe_error_api_connection_error');
-        } catch (\Stripe\Error\Base $e) {
-            $errorMessage = $e->getMessage();
+            auth()->user()->account
+                ->subscribe($request->input('stripeToken'), $request->input('plan'));
+        } catch (StripeException $e) {
+            return back()
+                ->withInput()
+                ->withErrors($e->getMessage());
         }
 
-        return back()
-            ->withInput()
-            ->withErrors($errorMessage);
+        return redirect()->route('settings.subscriptions.upgrade.success');
     }
 
     /**