+++ tests/Feature/RegisterTest.php
@@ -27,7 +27,6 @@ class RegisterTest extends FeatureTestCase
 
         $response = $this->post('/register', $params);
 
-        $response->dump();
         $response->assertStatus(302);
         $response->assertRedirect('/dashboard');